fix: Check errors when parsing numbers

+++ dreal/solver/context.cc
@@ -422,7 +422,7 @@ void Context::Minimize(const vector<Expression>& functions) {
 
 void Context::Maximize(const Expression& f) { impl_->Minimize({-f}); }
 
-void Context::Pop(int n) {
+void Context::Pop(long n) {
   DREAL_LOG_DEBUG("Context::Pop({})", n);
   if (n <= 0) {
     throw DREAL_RUNTIME_ERROR(
@@ -433,7 +433,7 @@ void Context::Pop(int n) {
   }
 }
 
-void Context::Push(int n) {
+void Context::Push(long n) {
   DREAL_LOG_DEBUG("Context::Push({})", n);
   if (n <= 0) {
     throw DREAL_RUNTIME_ERROR(
